cores: control: Implement LSM6DSO interface
This comit enables accelerometer and gyroscope data to be
collected from an LSM6DSO over I2C. This functionality can
be enablked by setting `CONFIG_APP_USE_IMU=y`.

This commit is the first of two, which aim to add support for
the IMU unit (LSM6DSO and LIS3MDL). The architecture follows
the workqueue-poll paradigm designed in the previous commit.