lib: heartbeat: Implement heartbeat timer library
This commit adds the implementation of a basic heartbeat timer.
The timer is packaged in a library to allow it to be used across
cores.

The timer has a configurable duration and toggles the gpio aliased
to heartbeat-gpio. This alias must appear in the devicetree, or the
library will not compile.