cores: control: Implement basic GNSS module
This commits add location tracking functionality to the control
core via GNSS. The module is based on the NCS Cellular: GNSS sample.
The control core application will now periodically try to acquire a
GNSS fix.

The module is implemented as a compile-time thread at present.
A future commit will refactor the module to use a runtime
thread instead. This will enable runtime creation and destruction
of the thread, making the module more lightweight and configurable.