cores: control: Implement sensors shell commands
This commit provides implementations for shell commands stubbed
in the previous commit. State related to the sensors module, i.e.
IMU, SHT40, is now viewable and configurable at runtime.

The main implication of this commit is that users may now update
the polling rate of the IMU and SHT40 at runtime.