lib: heartbeat: Disable heartbeat lib by default
Disable the hearbeat library by default to improve the testability
of libraries.

This change makes the heartbeat library an opt-in feature, allowing
for direct integration and unit testing of the library in a dedicated
application.

Furthermore, this uncouples libraries from each other, enabling
independent development and testing of future additions.