lib: icmp: Implement inflight message timeouts
Inflight messages will now be dropped from the inflight table
if a timeout duration is exceeded.

The implementation includes a new work item, which will be enqueued
to the system workqueue on the expiry of a k_timer. When the
timeout function is executed, the inflight table is scanned and
any messages older than CONFIG_ICMP_MAX_INFLIGHT_MSG_AGE
are dropped.