cores: interface: Add SDI-12 interface
Add an SDI-12 interface and a shell command to issue SDI-12
commands.

This first pass at the SDI-12 implementation is quite crude. It
lacks retries and could be improved. A future refactor should
move to a work item (co-routine) query method, with the SDI-12
driver operating as a state machine.